+++ GridKit/Model/PhasorDynamics/LoadZIP/LoadZIPDependencyTracking.cpp
@@ -1,0 +1,29 @@
+
+#include "LoadZIPImpl.hpp"
+
+namespace GridKit
+{
+  namespace PhasorDynamics
+  {
+    /**
+     * @brief Jacobian evaluation not implemented
+     *
+     * @tparam ScalarT - scalar data type
+     * @tparam IdxT    - matrix index data type
+     * @return int - error code, 0 = success
+     */
+    template <class ScalarT, typename IdxT>
+    int LoadZIP<ScalarT, IdxT>::evaluateJacobian()
+    {
+      Log::misc() << "Evaluate Jacobian for LoadZIP..." << std::endl;
+      Log::misc() << "Jacobian evaluation is not implemented!" << std::endl;
+
+      return 0;
+    }
+
+    // Available template instantiations
+    template class LoadZIP<DependencyTracking::Variable, long int>;
+    template class LoadZIP<DependencyTracking::Variable, size_t>;
+
+  } // namespace PhasorDynamics
+} // namespace GridKit