+++ GridKit/Model/PhasorDynamics/SystemModelData.hpp
@@ -12,6 +12,7 @@
 #include <GridKit/Model/PhasorDynamics/Exciter/IEEET1/Ieeet1Data.hpp>
 #include <GridKit/Model/PhasorDynamics/Governor/Tgov1/Tgov1Data.hpp>
 #include <GridKit/Model/PhasorDynamics/Load/LoadData.hpp>
+#include <GridKit/Model/PhasorDynamics/LoadZIP/LoadZIPData.hpp>
 #include <GridKit/Model/PhasorDynamics/SignalNode/SignalNodeData.hpp>
 #include <GridKit/Model/PhasorDynamics/SynchronousMachine/GENROUwS/GenrouData.hpp>
 #include <GridKit/Model/PhasorDynamics/SynchronousMachine/GenClassical/GenClassicalData.hpp>
@@ -37,6 +38,7 @@ namespace GridKit
       using GenrouDataT       = GenrouData<RealT, IdxT>;
       using GenClassicalDataT = GenClassicalData<RealT, IdxT>;
       using LoadDataT         = LoadData<RealT, IdxT>;
+      using LoadZIPDataT      = LoadZIPData<RealT, IdxT>;
       using SignalDataT       = SignalNodeData<RealT, IdxT>;
       using MonitorSinkSpec   = Model::VariableMonitorBase::SinkSpec;
 
@@ -83,6 +85,7 @@ namespace GridKit
       std::vector<GenrouDataT>       genrou;       ///< GENROU instances within the model
       std::vector<GenClassicalDataT> genclassical; ///< Classical generator instances within the model
       std::vector<LoadDataT>         load;         ///< Loads within the model
+      std::vector<LoadZIPDataT>      loadzip;      ///< Loads within the model
       std::vector<Tgov1DataT>        gov;          ///< Governors within the model
       std::vector<Ieeet1DataT>       exciter;      ///< Exciters within the model
       std::vector<SignalDataT>       signal;       ///< Signal nodes