+++ tests/UnitTests/PhasorDynamics/runLoadZIPTests.cpp
@@ -1,0 +1,18 @@
+#include "LoadZIPTests.hpp"
+
+int main()
+{
+  using namespace GridKit;
+  using namespace GridKit::Testing;
+
+  GridKit::Testing::TestingResults               result;
+  GridKit::Testing::LoadZIPTests<double, size_t> test;
+
+  result += test.constructor();
+  result += test.residual();
+#ifdef GRIDKIT_ENABLE_ENZYME
+  result += test.jacobian();
+#endif
+
+  return result.summary();
+}